Assignment 2
Write a C++ program that makes use of integer, floating point, and string variables. Specifically, your program does the following: (i) prompts the user to enter two integer values, and is able to perform basic math (+, -, *, /) on them, (ii) prompts the user to enter two floating point values, and is able to perform basic math on them, and (iii) prompts the user to enter two ‘string’ values (e.g., name and address) and is able to concatenate these two strings and print back the result.