Assignment 6
Your next assignment is to write a series of functions outside main, and utilize these functions by calling them from main(). 

The functions you must implement: 

A function that prints a simple text message on screen, such as "Hope you're having a good day". The signature for this function is: void printGreeting()
A function that accepts two floating point values as input (x and y), and prints the result of x/y. If the denominator (y) is zero, the function prints "Cannot perform division". Signature for this function: void divideNumbers(float x, float y)
A function that returns the factorial of an input integer value (num). If the input number is negative, the function returns 0. If the value is non-negative, the function computes the factorial and returns the value. Signature of this function: int factorial(int num) 
A function that accepts a positive integer as input (N), and prints the first N terms of the Fibonacci series. For example, for an input of N = 10, the function prints "0, 1, 1, 2, 3, 5, 8, 13, 21, 34". Signature for this function: void fibonacci(int N) 
[Final Step]: Once you have coded down these functions outside your main, go ahead and call these functions with meaningful parameters as input from within main(). If the function returns a value, capture and print that response in main.