Assignment 8
Your next assignment is to simulate basic actions of a fairly realistic “Vending Machine”. You must create a new class called “VendingMachine”, and perform the following using a variable (object) of this type in your main() function. 

Your VendingMachine serves four or more items to customers (e.g., Sprite, Water, etc.,). For each type, there is an initial stock of items (e.g., 10 cans of Sprite), as well as a price per item (e.g., Sprite is $1.25).  

Your program allows users to: 

 select an item from a simple menu (e.g., “Press 1 for Sprite”, “Press 2 for water” etc.,)

Enter a dollar amount (e.g, “Insert cash $5, $10, or $20”)

Your vending machine prints a success message on screen, as well as the correct change amount (e.g, User inserts $5 for Sprite costing $1.25, the correct change amount is $3.75)

If the machine runs out of stock for any item, it refunds the full amount back to the user.

If the machine is out of stock, there is an "admin" function to restock.