Assignment 9
Your final assignment is to simulate the actions of a realistic ATM machine. You must create a new “class” called ATM machine, and be able to create a variable (object) of this type in your main() function. Using this object, perform all the following functions: 

Initializes a bank account to $1000 (Hint: use the constructor) 
Supports a function to make a deposit. Deposits can be cash or check. Cash deposits can be any arbitrary value less than $100, while check deposits are limitless. 
Supports a function to withdraw cash. The maximum withdrawal amount is $200, and withdrawal amounts must always be a multiple of 20. 
Supports a simple function check balance, which prints the current account balance on screen.